DellEMC/benchmarks/ssd/implementation/mxnet/tests/requirements.txt
Jinja2==2.11.1
numpy==1.18.1
pandas==1.0.1
plotly==4.5.0
scipy==1.4.1
gitpython==3.1.0
